Generate and propagate Monte Carlo Simulations based on a template input file.

TBD

Usage

fct_combine_mcs_cstock(.ad, .cs, .usr)

Arguments

.ad

Activity Data input table for the shiny app (AD_lu_transitions)

.cs

Carbon Stock input table for the shiny app (c_stocks)

.usr

User inputs' table for the shiny app (user_inputs). Contains the number of iterations of the MCS, carbon fraction if needed and if truncated PDFs should be used when necessary.

Value

A data frame with Monte Carlo simulations of CO2 emissions for each land use transition, REDD+ activity or emission reductions level.

Examples

library(readxl)
library(dplyr)
library(mocaredd)

path <- system.file("extdata/example1-4pools.xlsx", package = "mocaredd")

cs <- read_xlsx(path = path, sheet = "c_stocks", na = "NA")
ad <- read_xlsx(path = path, sheet = "AD_lu_transitions", na = "NA")
usr <- read_xlsx(path = path, sheet = "user_inputs", na = "NA")

res <- fct_combine_mcs_cstock(.ad = ad, .cs = cs, .usr = usr)
res |> filter(sim_no == 1)
#> # A tibble: 18 × 11
#>    sim_no c_period lu_id     C_all C_form   AGB   BGB    DW    LI   ALL DG_ratio
#>     <int> <chr>    <chr>     <dbl> <chr>  <dbl> <dbl> <dbl> <dbl> <dbl>    <dbl>
#>  1      1 ALL      open      65.3  AGB +…  29.6  10.8  22.3  2.56 NA      NA    
#>  2      1 ALL      ev_wet_… 121.   AGB +…  95.9  11.9  10.0  2.89 NA      NA    
#>  3      1 ALL      ev_mois… 307.   AGB +… 253.   28.6  22.3  3.02 NA      NA    
#>  4      1 ALL      md_mois… 118.   AGB +…  40.8  26.0  48.5  2.53 NA      NA    
#>  5      1 ALL      md_mois… 172.   AGB +… 110.   21.1  38.4  2.39 NA      NA    
#>  6      1 ALL      ev_upla… 134.   AGB +…  67.3  27.5  37.9  1.43 NA      NA    
#>  7      1 ALL      postdef…  18.9  ALL     NA    NA    NA   NA    18.9    NA    
#>  8      1 ALL      postdef…   6.96 ALL     NA    NA    NA   NA     6.96   NA    
#>  9      1 ALL      postdef…  21.0  ALL     NA    NA    NA   NA    21.0    NA    
#> 10      1 ALL      postdef…   4.90 ALL     NA    NA    NA   NA     4.90   NA    
#> 11      1 ALL      postdef…  18.0  ALL     NA    NA    NA   NA    18.0    NA    
#> 12      1 ALL      postdef…   6.63 ALL     NA    NA    NA   NA     6.63   NA    
#> 13      1 ALL      open_deg  36.5  DG_ra…  NA    NA    NA   NA    NA       0.582
#> 14      1 ALL      ev_wet_…  91.6  DG_ra…  NA    NA    NA   NA    NA       0.777
#> 15      1 ALL      ev_mois… 232.   DG_ra…  NA    NA    NA   NA    NA       0.763
#> 16      1 ALL      md_mois…  92.3  DG_ra…  NA    NA    NA   NA    NA       0.801
#> 17      1 ALL      md_mois… 121.   DG_ra…  NA    NA    NA   NA    NA       0.711
#> 18      1 ALL      ev_upla…  95.9  DG_ra…  NA    NA    NA   NA    NA       0.722