Generate and propagate Monte Carlo Simulations based on a template input file.

TBD

Usage

fct_combine_mcs_cstock(.ad, .cs, .usr)

Arguments

.ad

Activity Data input table for the shiny app (AD_lu_transitions)

.cs

Carbon Stock input table for the shiny app (c_stocks)

.usr

User inputs' table for the shiny app (user_inputs). Contains the number of iterations of the MCS, carbon fraction if needed and if truncated PDFs should be used when necessary.

Value

A data frame with Monte Carlo simulations of CO2 emissions for each land use transition, REDD+ activity or emission reductions level.

Examples

library(readxl)
library(dplyr)
library(mocaredd)

cs <- read_xlsx(
  system.file("extdata/example1.xlsx", package = "mocaredd"),
  sheet = "c_stocks",
  na = "NA"
  )
ad <- read_xlsx(
  system.file("extdata/example1.xlsx", package = "mocaredd"),
  sheet = "AD_lu_transitions",
  na = "NA"
  )
usr <- read_xlsx(
  system.file("extdata/example1.xlsx", package = "mocaredd"),
  sheet = "user_inputs",
  na = "NA"
  )

res <- fct_combine_mcs_cstock(.ad = ad, .cs = cs, .usr = usr)
res |> filter(sim_no == 1)
#> # A tibble: 18 × 11
#>    sim_no c_period lu_id     C_all C_form   AGB   BGB    DW    LI   ALL DG_ratio
#>     <int> <chr>    <chr>     <dbl> <chr>  <dbl> <dbl> <dbl> <dbl> <dbl>    <dbl>
#>  1      1 ALL      open    239.    (AGB …  30.2  11.4 20.6   3.00 NA      NA    
#>  2      1 ALL      ev_wet… 573.    (AGB … 102.   10.8 40.2   3.16 NA      NA    
#>  3      1 ALL      ev_moi… 835.    (AGB … 191.   27.5  7.05  2.64 NA      NA    
#>  4      1 ALL      md_moi… 629.    (AGB …  51.4  24.3 92.9   2.85 NA      NA    
#>  5      1 ALL      md_moi… 503.    (AGB …  78.5  21.3 35.7   1.76 NA      NA    
#>  6      1 ALL      ev_upl… 464.    (AGB …  72.3  28.5 24.1   1.55 NA      NA    
#>  7      1 ALL      postde…  90.6   ALL *…  NA    NA   NA    NA    24.7    NA    
#>  8      1 ALL      postde…  44.3   ALL *…  NA    NA   NA    NA    12.1    NA    
#>  9      1 ALL      postde…  87.6   ALL *…  NA    NA   NA    NA    23.9    NA    
#> 10      1 ALL      postde…  39.8   ALL *…  NA    NA   NA    NA    10.8    NA    
#> 11      1 ALL      postde…  19.2   ALL *…  NA    NA   NA    NA     5.25   NA    
#> 12      1 ALL      postde… 102.    ALL *…  NA    NA   NA    NA    27.9    NA    
#> 13      1 ALL      dg_open   0.440 DG_ra…  NA    NA   NA    NA    NA       0.440
#> 14      1 ALL      dg_ev_…   0.654 DG_ra…  NA    NA   NA    NA    NA       0.654
#> 15      1 ALL      dg_ev_…   0.698 DG_ra…  NA    NA   NA    NA    NA       0.698
#> 16      1 ALL      dg_md_…   0.656 DG_ra…  NA    NA   NA    NA    NA       0.656
#> 17      1 ALL      dg_md_…   0.695 DG_ra…  NA    NA   NA    NA    NA       0.695
#> 18      1 ALL      dg_ev_…   0.687 DG_ra…  NA    NA   NA    NA    NA       0.687