Check that the app's the input data has matching land uses and correct carbon pools and associated factors.

The expected input carbon stock table for the tool contains carbon values of different carbon pools coded: - AGB for aboveground biomass, - BGB or RS (Root-to-Shoot ratio) for the belowground biomass, - DW for deadwood, - LI for litter and - SOC for soil organic carbon. All pools can be expressed in tons of Carbon (C) or CO2. AGB and BGB can also be expressed in ton of dry matter (DM), then a carbon fraction (CF) is also needed.

Usage

fct_make_formula(.c_check, .c_unit)

Arguments

.c_check

output of fct_check_pool()

.c_unit

"DM" or "C",

Value

A character value with the formula for calculating total carbon stock for a specific land use.

Examples

library(mocaredd)
library(readxl)
library(dplyr)

path <- system.file("extdata/example1-4pools.xlsx", package = "mocaredd")

cs <- read_xlsx(path = path, sheet = "c_stocks", na = "NA")

c_lu  <- cs |> filter(lu_id == "ev_wet_closed")
#> Error in filter(cs, lu_id == "ev_wet_closed"): ℹ In argument: `lu_id == "ev_wet_closed"`.
#> Caused by error:
#> ! object 'lu_id' not found

c_check <- fct_check_pool(.c_lu = c_lu, .c_unit = "C", .c_fraction = NA)
#> Error in fct_check_pool(.c_lu = c_lu, .c_unit = "C", .c_fraction = NA): unused argument (.c_lu = c_lu)

fct_make_formula(.c_check = c_check, .c_unit = "C")
#> Error: object 'c_check' not found